test: finish test package cleanup
The shared Git history fixture created a fresh writable global config for every command. Git-heavy packages paid that repeated setup cost even though each test process already owns an isolated config.

Reuse the process-wide isolated runner for ordinary fixture commands. Keep runtime attachment tests in their own file so the basic route tests stay focused on HTTP behavior.
