+++ internal/testutil/gitfixture/history.go
@@ -19,30 +19,36 @@ import (
 func DivergenceWorktree(t *testing.T) string {
 	t.Helper()
 	require := require.New(t)
+	runner := gitsafe.Runner().WithConfig("init.defaultBranch", "main")
+	run := func(dir string, args ...string) {
+		t.Helper()
+		out, stderr, err := runner.Run(t.Context(), dir, nil, args...)
+		require.NoError(err, "git %v failed: %s%s", args, out, stderr)
+	}
 
 	root := t.TempDir()
 	remote := filepath.Join(root, "remote.git")
 	work := filepath.Join(root, "work")
-	Run(t, root, "init", "--bare", "--initial-branch=main", remote)
-	Run(t, root, "clone", remote, work)
-	Run(t, work, "config", "user.email", "t@test.com")
-	Run(t, work, "config", "user.name", "Test")
+	run(root, "init", "--bare", "--initial-branch=main", remote)
+	run(root, "clone", remote, work)
+	run(work, "config", "user.email", "t@test.com")
+	run(work, "config", "user.name", "Test")
 	require.NoError(os.WriteFile(filepath.Join(work, "base.txt"), []byte("base\n"), 0o644))
-	Run(t, work, "add", ".")
-	Run(t, work, "commit", "-m", "base")
-	Run(t, work, "push", "origin", "main")
-	Run(t, work, "checkout", "-b", "feature")
+	run(work, "add", ".")
+	run(work, "commit", "-m", "base")
+	run(work, "push", "origin", "main")
+	run(work, "checkout", "-b", "feature")
 	require.NoError(os.WriteFile(filepath.Join(work, "f.txt"), []byte("f1\n"), 0o644))
-	Run(t, work, "add", ".")
-	Run(t, work, "commit", "-m", "feature 1")
-	Run(t, work, "push", "-u", "origin", "feature")
+	run(work, "add", ".")
+	run(work, "commit", "-m", "feature 1")
+	run(work, "push", "-u", "origin", "feature")
 	return work
 }
 
-// Run executes Git with a per-test isolated configuration.
+// Run executes Git with the package test process's isolated configuration.
 func Run(t *testing.T, dir string, args ...string) []byte {
 	t.Helper()
-	runner := gitsafe.MutableRunner(t).WithConfig("init.defaultBranch", "main")
+	runner := gitsafe.Runner().WithConfig("init.defaultBranch", "main")
 	out, stderr, err := runner.Run(t.Context(), dir, nil, args...)
 	require.NoError(t, err, "git %v failed: %s%s", args, out, stderr)
 	return out
